x/cctp/keeper/grpc_query_next_available_nonce.go
package keeper

import (
	"context"

	"github.com/circlefin/noble-cctp/x/cctp/types"
	sdk "github.com/cosmos/cosmos-sdk/types"
	"google.golang.org/grpc/codes"
	"google.golang.org/grpc/status"
)

func (k Keeper) NextAvailableNonce(c context.Context, req *types.QueryGetNextAvailableNonceRequest) (*types.QueryGetNextAvailableNonceResponse, error) {
	if req == nil {
		return nil, status.Error(codes.InvalidArgument, "invalid request")
	}
	ctx := sdk.UnwrapSDKContext(c)

	val, found := k.GetNextAvailableNonce(ctx)
	if !found {
		return nil, status.Error(codes.NotFound, "not found")
	}

	return &types.QueryGetNextAvailableNonceResponse{Nonce: val}, nil
}
